ManyToMany role

+++ Backend/src/main/java/com/myproject/library/entity/Role.java
@@ -12,11 +12,12 @@ import lombok.NoArgsConstructor;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(unique = true, nullable = false)
     private String roleName;
 
     public static String ADMIN = "ADMIN";
     public static String USER = "USER";
+
 }
